Move tests/common/plugins/tacacs.py to tests/tacacs (#4476)
Description of PR
1. File tests/common/plugins/tacacs.py is just for tacacs testing. Need to move it to test specific folder tests/tacacs. It should not be under tests/common folder.
2. Suggest to rename the 'test_' fixtures to 'check_' in tests/common/plugins/tacacs.py. If it just a fixture to be used by a test, we should not name it 'test_*'.

What is the motivation for this PR?
1. File tests/common/plugins/tacacs.py is just for tacacs testing. Need to move it to test specific folder tests/tacacs. It should not be under tests/common folder.
2. Suggest to rename the 'test_' fixtures to 'check_' in tests/common/plugins/tacacs.py. If it just a fixture to be used by a test, we should not name it 'test_*'.

How did you do it?
1. Move codes in file tests/common/plugins/tacacs.py into tests/tacacs/utils.py, and add a new file tests/tacacs/conftest.py in order to use the fixtures. At the same time, modify the test/conftest.py to delete the unnecessary information.
2. Rename the 'test_' fixtures to 'check_' in tests/tacacs/conftest.py, and modify the functions who use these fixtures.

How did you verify/test it?
I test it by running the test cases under tests/tacacs.
