+++ tests/tacacs/test_rw_user.py
@@ -11,7 +11,7 @@ pytestmark = [
 ]
 
 
-def test_rw_user(localhost, duthosts, enum_rand_one_per_hwsku_hostname, creds_all_duts, test_tacacs):
+def test_rw_user(localhost, duthosts, enum_rand_one_per_hwsku_hostname, creds_all_duts, check_tacacs):
     """test tacacs rw user
     """
     duthost = duthosts[enum_rand_one_per_hwsku_hostname]
@@ -21,7 +21,7 @@ def test_rw_user(localhost, duthosts, enum_rand_one_per_hwsku_hostname, creds_al
 
     check_output(res, 'testadmin', 'remote_user_su')
 
-def test_rw_user_ipv6(localhost, duthosts, enum_rand_one_per_hwsku_hostname, creds_all_duts, test_tacacs_v6):
+def test_rw_user_ipv6(localhost, duthosts, enum_rand_one_per_hwsku_hostname, creds_all_duts, check_tacacs_v6):
     """test tacacs rw user
     """
     duthost = duthosts[enum_rand_one_per_hwsku_hostname]
